# api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Sendit Documentation",
...
        "contact": {
            "name": "Sendit Support",
            "email": "contact@sendit.ma"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://app.sendit.ma/api/v1/",
            "description": "Sendit"
        }
    ],
    "paths": {
...
        "/districts": {
            "get": {
                "tags": [
                    "Villes"
                ],
                "description": "Liste des villes",
                "operationId": "e18e02f18e7355d8aa4ca8186f88c2cb",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "querystring",
                        "in": "query",
                        "description": "Recherche par nom de la ville.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "rabat"
                        }
                    },
                    {
                        "name": "pickup-district",
                        "in": "query",
                        "description": "Identifiant de la ville de ramassage. Les frais et délais de livraison affichés sont ajustés selon la ville de ramassage sélectionnée. Par défaut l'identifiant de Casablanca c'est id n° 46.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 46
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success, Retourner une liste des villes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des villes."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/District"
                                            }
                                        },
                                        "total": {
                                            "description": "Le nombre total d'éléments dans la collection.",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "per_page": {
                                            "description": "Le nombre d'éléments affichés par page.",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "current_page": {
                                            "description": "Le numéro de la page actuellement affichée.",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "description": "Le numéro de la dernière page de la collection.",
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "next_page_url": {
                                            "description": "L'URL de la page suivante de la collection. Si la page actuelle est la dernière, cette valeur peut être null.",
                                            "type": "string",
                                            "example": "http://example.com/api/v1/districts?page=2"
                                        },
                                        "prev_page_url": {
                                            "description": "L'URL de la page précédente de la collection. Si la page actuelle est la première, cette valeur peut être null.",
                                            "type": "string",
                                            "example": "http://example.com/api/v1/districts?page=1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Vous n'avez pas les autorisations nécessaires"
                    },
                    "404": {
                        "description": "Page non trouvée"
                    },
                    "500": {
                        "description": "Une erreur s'est produite côté serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Liste des produits",
                "description": "Retourne la liste des produits",
                "operationId": "24040067e52a418977eccdf09f28c3fd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "querystring",
                        "in": "query",
                        "description": "Recherche par code,nom et référence du produit",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductList"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Vous n'avez pas les autorisations nécessaires"
                    },
                    "404": {
                        "description": "Page non trouvée"
                    },
                    "500": {
                        "description": "Une erreur s'est produite côté serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un nouveau produit",
                "description": "Crée un produit en fournissant les informations nécessaires",
                "operationId": "bf0f187b6a66fd5d6c45801fd6e407be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reference",
                                    "name",
                                    "price",
                                    "category",
                                    "active",
                                    "image"
                                ],
                                "properties": {
                                    "reference": {
                                        "description": "Référence unique du produit",
                                        "type": "string",
                                        "example": "ref_X"
                                    },
                                    "name": {
                                        "description": "Nom du produit",
                                        "type": "string",
                                        "example": "Produit X"
                                    },
                                    "price": {
                                        "description": "Prix du produit",
                                        "type": "number",
                                        "format": "float",
                                        "example": 199
                                    },
                                    "category": {
                                        "description": "Catégorie du produit (valeurs autorisées)",
                                        "type": "string",
                                        "enum": [
                                            "Électronique | الإلكترونيات",
                                            "Vêtements | الملابس",
                                            "Accessoires de mode | إكسسوارات الموضة",
                                            "Beauté et santé | الجمال والصحة",
                                            "Maison et jardin | المنزل والحديقة",
                                            "Jeux et jouets | الألعاب والدمى",
                                            "Livres et papeterie | الكتب والقرطاسية",
                                            "Sports et loisirs | الرياضة والترفيه",
                                            "Produits pour animaux | منتجات للحيوانات الأليفة",
                                            "Informatique | الحواسيب",
                                            "Téléphones et accessoires | الهواتف والملحقات",
                                            "Articles de voyage | مستلزمات السفر",
                                            "Bijoux | المجوهرات",
                                            "Musique et instruments | الموسيقى والآلات الموسيقية",
                                            "Articles pour bébés | مستلزمات الأطفال",
                                            "Voitures et accessoires | السيارات وملحقاتها",
                                            "Articles de bureau | لوازم المكتب",
                                            "Autres | الأخرى"
                                        ],
                                        "example": "Accessoires de mode | إكسسوارات الموضة"
                                    },
                                    "active": {
                                        "description": "Statut du produit (1=actif,0=inactif)",
                                        "type": "boolean",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "Image encodée en Base64 avec Data URI (data:image/png;base64,...)",
                                        "type": "string",
                                        "format": "byte",
                                        "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAQAAAAEACAYAAABccqhmAAAABHNCSVQICAgIfAhkiAAAA..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDetail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Les données fournies sont invalides"
                    },
                    "403": {
                        "description": "Vous n'avez pas les autorisations nécessaires"
                    },
                    "500": {
                        "description": "Une erreur s'est produite côté serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "District": {
                "title": "Ville",
                "properties": {
                    "id": {
                        "description": "Identifiant de la ville.",
                        "type": "string",
                        "example": "1"
                    },
                    "ville": {
                        "description": "Nom de la ville.",
                        "type": "string",
                        "example": "Ville X"
                    },
                    "name": {
                        "description": "Nom de la quartier/ville ",
                        "type": "string",
                        "example": "Quartier/ville X"
                    },
                    "arabic_name": {
                        "description": "Nom de la quartier/ville en arabe",
                        "type": "string",
                        "example": ""
                    },
                    "price": {
                        "description": "Frais de livraison de la ville",
                        "type": "number",
                        "format": "float",
                        "example": "X"
                    },
                    "delais": {
                        "type": "string",
                        "example": "24h"
                    },
                    "pickup_district": {
                        "description": "Indique si la ville est une ville de ramassage. Les valeurs possibles sont 0 pour non et 1 pour oui.",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "ProductList": {
                "title": "Liste des produits",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "message": {
                        "type": "string",
                        "example": "Liste des Produits."
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductData"
                        }
                    },
                    "total": {
                        "description": "Le nombre total d'éléments dans la collection.",
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "description": "Le nombre d'éléments affichés par page.",
                        "type": "integer",
                        "example": 10
                    },
                    "current_page": {
                        "description": "Le numéro de la page actuellement affichée.",
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "description": "Le numéro de la dernière page de la collection.",
                        "type": "integer",
                        "example": 10
                    },
                    "next_page_url": {
                        "description": "L'URL de la page suivante de la collection. Si la page actuelle est la dernière, cette valeur peut être null.",
                        "type": "string",
                        "example": "http://example.com/api/v1/districts?page=2"
                    },
                    "prev_page_url": {
                        "description": "L'URL de la page précédente de la collection. Si la page actuelle est la première, cette valeur peut être null.",
                        "type": "string",
                        "example": "http://example.com/api/v1/districts?page=1"
                    }
                },
                "type": "object"
            },
            "ProductData": {
                "title": "Informations sur le produit",
                "properties": {
                    "id": {
                        "description": "Identifiant du produit.",
                        "type": "integer",
                        "example": "1"
                    },
                    "reference": {
                        "description": "Référence du produit.",
                        "type": "string",
                        "example": "ref_X"
                    },
                    "code": {
                        "description": "Code du produit.",
                        "type": "string",
                        "example": "PXXXXXX"
                    },
                    "name": {
                        "description": "Nom du produit.",
                        "type": "string",
                        "example": "Produit X"
                    },
                    "price": {
                        "description": "Prix du produit.",
                        "type": "number",
                        "format": "float",
                        "example": "X"
                    },
                    "quantity": {
                        "description": "Quantité en stock.",
                        "type": "integer",
                        "example": "100"
                    },
                    "active": {
                        "description": "État du produit.",
                        "type": "integer",
                        "example": "1"
                    },
                    "quantity_to_prepare": {
                        "description": "Quantité à préparer.",
                        "type": "integer",
                        "example": "10"
                    }
                },
                "type": "object"
            },
            "ProductDetail": {
                "title": "Détail du produit",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "message": {
                        "type": "string",
                        "example": "Détail du produit."
                    },
                    "data": {
                        "$ref": "#/components/schemas/ProductDetailData"
                    }
                },
                "type": "object"
            },
            "ProductDetailData": {
                "title": "Détail du produit - informations",
                "properties": {
                    "id": {
                        "description": "Identifiant du produit.",
                        "type": "integer",
                        "example": 1
                    },
                    "reference": {
                        "description": "Référence du produit.",
                        "type": "string",
                        "example": "ref_X"
                    },
                    "code": {
                        "description": "Code du produit",
                        "type": "string",
                        "example": "PXXXXX"
                    },
                    "name": {
                        "description": "Nom du produit",
                        "type": "string",
                        "example": "Produit X"
                    },
                    "price": {
                        "description": "Prix du produit",
                        "type": "number",
                        "format": "float",
                        "example": 199
                    },
                    "quantity": {
                        "description": "Quantité du produit",
                        "type": "integer",
                        "example": 10
                    },
                    "active": {
                        "description": "Statut du produit (1=actif,0=inactif)",
                        "type": "integer",
                        "example": 1
                    },
                    "quantity_to_prepare": {
                        "description": "Quantité à préparer du produit",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}